+++ src/rules/newline-after-import.js
@@ -124,7 +124,7 @@ module.exports = {
       }
     }
 
-    function commentAfterImport(node, nextComment) {
+    function commentAfterImport(node, nextComment, type) {
       const lineDifference = getLineDifference(node, nextComment);
       const EXPECTED_LINE_DIFFERENCE = options.count + 1;
 
@@ -140,7 +140,7 @@ module.exports = {
             line: node.loc.end.line,
             column,
           },
-          message: `Expected ${options.count} empty line${options.count > 1 ? 's' : ''} after import statement not followed by another import.`,
+          message: `Expected ${options.count} empty line${options.count > 1 ? 's' : ''} after ${type} statement not followed by another ${type}.`,
           fix: options.exactCount && EXPECTED_LINE_DIFFERENCE < lineDifference ? undefined : (fixer) => fixer.insertTextAfter(
             node,
             '\n'.repeat(EXPECTED_LINE_DIFFERENCE - lineDifference),
@@ -178,7 +178,7 @@ module.exports = {
       }
 
       if (nextComment && typeof nextComment !== 'undefined') {
-        commentAfterImport(node, nextComment);
+        commentAfterImport(node, nextComment, 'import');
       } else if (nextNode && nextNode.type !== 'ImportDeclaration' && (nextNode.type !== 'TSImportEqualsDeclaration' || nextNode.isExport)) {
         checkForNewLine(node, nextNode, 'import');
       }
@@ -215,8 +215,18 @@ module.exports = {
               || !containsNodeOrEqual(nextStatement, nextRequireCall)
             )
           ) {
+            let nextComment;
+            if (typeof statementWithRequireCall.parent.comments !== 'undefined' && options.considerComments) {
+              const endLine = node.loc.end.line;
+              nextComment = statementWithRequireCall.parent.comments.find((o) => o.loc.start.line >= endLine && o.loc.start.line <= endLine + options.count + 1);
+            }
 
-            checkForNewLine(statementWithRequireCall, nextStatement, 'require');
+            if (nextComment && typeof nextComment !== 'undefined') {
+
+              commentAfterImport(statementWithRequireCall, nextComment, 'require');
+            } else {
+              checkForNewLine(statementWithRequireCall, nextStatement, 'require');
+            }
           }
         });
       },